+++ .github/CODEOWNERS
@@ -38,7 +38,9 @@ crates/common_utils/src/events/ @juspay/hyperswitch-analytics
 crates/common_utils/src/events.rs @juspay/hyperswitch-analytics
 crates/analytics/ @juspay/hyperswitch-analytics
 
+scripts/add_connector.sh @juspay/hyperswitch-connector
 connector-template/ @juspay/hyperswitch-connector
+crates/router/src/core/connector_validation.rs @juspay/hyperswitch-connector
 crates/router/src/connector/ @juspay/hyperswitch-connector
 crates/router/tests/connectors/ @juspay/hyperswitch-connector
 crates/test_utils/tests/connectors/ @juspay/hyperswitch-connector
